test
Worked on this a little bit during my free time at work, so far I've got a couple of things down. I've created the towers in which my disks will rest on, and have created the first disk. We need to implement the rest of them and a method for them to move around and ensure that a larger disk cannot lie on top of a smaller one. So far, I'm feeling good about this, I think it should go smoothly for the most part.

This is also a good way for me to start to get to know how to use GitHub in case I'll ever need it for anything.