V3
I've done a lot of work today, the program definitely has a long way to go before it is perfect or anywhere close to what I had envisioned it to be, however, at least I've got something up and running. 

You are now able to "solve" the puzzle. All disks to the puzzle are now interchangeable within towers but there is definitely still a multitute of bugs. No, reset button for starting the puzzle over, no counter of times object is moved, and many many many bugs when stacking the disks.